+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/legacy/types/javascript_channel.dart
