+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/legacy/types/types.dart
@@ -2,12 +2,13 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+export 'auto_media_playback_policy.dart';
+export 'creation_params.dart';
+export 'javascript_channel.dart';
 export 'javascript_message.dart';
 export 'javascript_mode.dart';
-export 'load_request_params.dart';
-export 'platform_navigation_delegate_creation_params.dart';
-export 'platform_webview_controller_creation_params.dart';
-export 'platform_webview_cookie_manager_creation_params.dart';
-export 'platform_webview_widget_creation_params.dart';
 export 'web_resource_error.dart';
+export 'web_resource_error_type.dart';
+export 'web_settings.dart';
 export 'webview_cookie.dart';
+export 'webview_request.dart';
