+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/types/webview_cookie.dart
@@ -2,15 +2,18 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-/// A cookie that can be set globally for all web views
-/// using [WebViewCookieManagerPlatform].
+import 'package:flutter/foundation.dart';
+
+/// A cookie that can be set globally for all web views using [WebViewCookieManagerPlatform].
+@immutable
 class WebViewCookie {
-  /// Constructs a new [WebViewCookie].
-  const WebViewCookie(
-      {required this.name,
-      required this.value,
-      required this.domain,
-      this.path = '/'});
+  /// Creates a new [WebViewCookieDelegate]
+  const WebViewCookie({
+    required this.name,
+    required this.value,
+    required this.domain,
+    this.path = '/',
+  });
 
   /// The cookie-name of the cookie.
   ///
@@ -30,20 +33,9 @@ class WebViewCookie {
   /// https://datatracker.ietf.org/doc/html/draft-ietf-httpbis-rfc6265bis-02#section-4.1.1
   final String domain;
 
-  /// The path-value of the cookie.
-  /// Is set to `/` in the constructor by default.
+  /// The path-value of the cookie, set to `/` by default.
   ///
   /// Its value should match "path-value" in RFC6265bis:
   /// https://datatracker.ietf.org/doc/html/draft-ietf-httpbis-rfc6265bis-02#section-4.1.1
   final String path;
-
-  /// Serializes the [WebViewCookie] to a Map<String, String>.
-  Map<String, String> toJson() {
-    return <String, String>{
-      'name': name,
-      'value': value,
-      'domain': domain,
-      'path': path
-    };
-  }
 }
