--- packages/webview_flutter/webview_flutter_platform_interface/lib/v4/webview_flutter_platform_interface.dart
@@ -1,10 +1,0 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-export 'src/platform_navigation_delegate.dart';
-export 'src/platform_webview_controller.dart';
-export 'src/platform_webview_cookie_manager.dart';
-export 'src/platform_webview_widget.dart';
-export 'src/types/types.dart';
-export 'src/webview_platform.dart';
