+++ packages/webview_flutter/webview_flutter_platform_interface/lib/webview_flutter_platform_interface_legacy.dart
@@ -1,0 +1,6 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+export 'src/legacy/platform_interface/platform_interface.dart';
+export 'src/legacy/types/types.dart';
