+++ packages/webview_flutter/webview_flutter_platform_interface/test/legacy/types/webview_cookie_test.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:webview_flutter_platform_interface/src/types/types.dart';
+import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface_legacy.dart';
 
 void main() {
   test('WebViewCookie should serialize correctly', () {
