+++ packages/webview_flutter/webview_flutter_platform_interface/test/legacy/types/webview_request_test.dart
@@ -4,7 +4,7 @@
 
 import 'dart:typed_data';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:webview_flutter_platform_interface/src/types/types.dart';
+import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface_legacy.dart';
 
 void main() {
   test('WebViewRequestMethod should serialize correctly', () {
